requirements.txt
mlflow
statsmodels
pymc
streamlit
jupyter
pydantic
aiobotocore==2.7.0
aiohttp==3.9.0
aioitertools==0.11.0
aiosignal==1.3.1
botocore==1.31.64
frozenlist==1.4.0
fsspec==2023.10.0
gcsfs==2023.10.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
jmespath==1.0.1
linearmodels
multidict==6.0.4
numpyro
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
s3fs==2023.10.0
st-files-connection==0.1.0
st-pages==0.4.5
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.2